modified:   src/App.js 	deleted:    src/components/aboutus.jsx 	renamed:    src/components/aboutcard.jsx -> src/components/aboutus/aboutcard.jsx 	renamed:    src/components/advertiserDashboard/makeAd.jsx -> src/components/makeAdvertisement/makeAd.jsx 	renamed:    src/components/advertiserDashboard/makeAdComponents.jsx -> src/components/makeAdvertisement/makeAdComponents.jsx 	renamed:    src/components/advertiserDashboard/makeForm.jsx -> src/components/makeAdvertisement/makeForm.jsx 	renamed:    src/components/advertiserDashboard/makeFormComponents.jsx -> src/components/makeAdvertisement/makeFormComponents.jsx 	modified:   src/components/navbar.jsx 	modified:   src/pages/aboutus.jsx 	renamed:    src/pages/advertiserDashboard.jsx -> src/pages/makeAdvertisement.jsx
